fix schema tests
